findaton/requirements.txt
promptflow
promptflow-tools
promptflow-vectordb
pandas
azure-search-documents==11.4.0
openai==1.3.0
num2words
matplotlib
plotly
scipy
scikit-learn
tiktoken